pyproject.toml
[project]
name = "agent-interview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "aiosqlite",
    "google-genai",
    "groq",
    "pyyaml",
    "python-dotenv",
    "python-multipart",
]
